Fixed Small Bugs, Added Sorting to Results
Added the ability to sort results based on either date or price. Also
added the option to only view posts that have photos (since I know
that's important when I look for housing).

Should we want to add more ways to sort (based on number of rooms,
square footage, price per square foot etc.) we can just have Kelly add
these numbers to the Post object and it should be very very easy to
write the code for sorting.